added tests to increase test coverage logging bc integration tests
